image: move validate-seed related tests to validate_seed_test.go
